+++ hpeOneView/resources/activity/tasks.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2021] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,11 +21,13 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 from future import standard_library
+import json
 
 standard_library.install_aliases()
 
 
 from hpeOneView.resources.resource import Resource, ResourcePatchMixin
+from hpeOneView.exceptions import HPEOneViewException
 
 
 class Tasks(ResourcePatchMixin, Resource):
@@ -38,7 +40,7 @@ class Tasks(ResourcePatchMixin, Resource):
     def __init__(self, connection, data=None):
         super(Tasks, self).__init__(connection, data)
 
-    def get_all(self, start=0, count=-1, fields='', filter='', query='', sort='', view=''):
+    def get_all(self, start=0, count=-1, fields='', filter='', query='', sort='', view='', topCount=0, childLimit=0):
         """
         Gets all the tasks based upon filters provided.
 
@@ -68,9 +70,43 @@ class Tasks(ResourcePatchMixin, Resource):
                  Returns a specific subset of the attributes of the resource or collection, by specifying the name of a
                  predefined view. The default view is expand (show all attributes of the resource and all elements of
                  collections of resources).
+            childLimit:
+                 Total number of associated resources in an aggregated manner. Default value is 10.
+            topCount:
+                 Total number of immediate children the task should send back. Otherwise, the task sends back the
+                 aggregated view of the tree. Default value is 3.
 
         Returns:
             list: A list of tasks.
         """
         return self._helper.get_all(start=start, count=count, filter=filter, query=query, sort=sort, view=view,
-                                    fields=fields)
+                                    fields=fields, childLimit=childLimit, topCount=topCount)
+
+    def patch(self, uri, timeout=-1):
+        """
+        Sets the state of task to cancelling only if IsCancellable is set to true for the task and its children or
+        children are in terminal state.
+
+        Args:
+            uri: URI of task resource.
+            timeout: Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
+                in OneView; it just stops waiting for its completion.
+
+        Returns:
+            dict: Updated resource.
+        """
+        resp, body = self._connection.do_http('PATCH', path=uri, body=None)
+
+        if resp.status >= 400:
+            raise HPEOneViewException(body)
+        elif resp.status == 304:
+            if body and not isinstance(body, dict):
+                try:
+                    body = json.loads(body)
+                except Exception:
+                    pass
+        elif resp.status == 202:
+            task = self._connection.__get_task_from_response(resp, body)
+            return self._task_monitor.wait_for_task(task, timeout)
+
+        return body