+++ hpeOneView/resources/search/labels.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2021] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -25,10 +25,27 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpeOneView.resources.resource import ResourceClient
+from hpeOneView.resources.resource import Resource
+from hpeOneView.resources.search.index_resources import IndexResources
+
+category_list = ['connection-templates', 'drive-enclosures', 'ethernet-networks', 'enclosures', 'enclosure-groups',
+                 'fc-networks', 'fcoe-networks', 'firmware-bundles', 'hypervisor-cluster-profiles',
+                 'hypervisor-host-profiles', 'hypervisor-managers', 'interconnects', 'logical-enclosures',
+                 'logical-interconnect-groups', 'logical-interconnects', 'network-sets', 'os-deployment-plans',
+                 'sas-interconnects', 'sas-logical-interconnect-groups', 'sas-logical-interconnects',
+                 'sas-logical-jbods', 'scopes', 'server-hardware', 'server-profile-templates', 'server-profiles',
+                 'storage-pools', 'storage-volume-sets', 'storage-volume-templates', 'storage-volumes']
 
 
-class Labels(object):
+def get_resources_associated_with_label(connection, label):
+    index_resource = IndexResources(connection)
+    query_string = "labels='{}'".format(label)
+    all_index_resources = index_resource.get_all(category=category_list, query=query_string)
+    response = [dict_response['uri'] for dict_response in all_index_resources]
+    return response
+
+
+class Labels(Resource):
     """
     Labels API client.
 
@@ -37,46 +54,37 @@ class Labels(object):
     URI = '/rest/labels'
     RESOURCES_PATH = '/resources'
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
+    def __init__(self, connection, data=None):
+        super(Labels, self).__init__(connection, data)
 
-    def get_all(self, start=0, count=-1, filter='', sort=''):
+    def get_assigned_resources(self, name):
         """
-        Gets a list of labels based on optional sorting and filtering and is constrained by start
-        and count parameters.
+        Retrieves the assigned resource of a label by name.
 
         Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
+            name: name of the label.
 
         Returns:
-            list: A list of labels.
+          Dict:  Get assigned resources of specified label resource.
         """
-        return self._client.get_all(start=start, count=count, filter=filter, sort=sort)
+        response = super(Labels, self).get_by_name(name)
+        if response:
+            resources_list = get_resources_associated_with_label(self._connection, name)
+            assigned_resource = dict(name=name, assignedResourceUris=resources_list)
+        return assigned_resource
 
-    def get(self, id_or_uri):
+    def create(self, resource, timeout=-1):
         """
-        Gets a label by ID or URI.
+        Set all the labels for a resource.
 
         Args:
-            id_or_uri: Can be either the label ID or the label URI.
+            resource: The object containing the resource URI and a list of labels
 
         Returns:
-            dict: The label.
+            dict: Resource Labels
         """
-        return self._client.get(id_or_uri=id_or_uri)
+        uri = self.URI + self.RESOURCES_PATH
+        return super(Labels, self).create(resource, uri=uri, timeout=timeout)
 
     def get_by_resource(self, resource_uri):
         """
@@ -89,41 +97,24 @@ class Labels(object):
             dict: Resource Labels
         """
         uri = self.URI + self.RESOURCES_PATH + '/' + resource_uri
-        return self._client.get(id_or_uri=uri)
+        return super(Labels, self).get_by_uri(uri)
 
-    def create(self, resource):
+    def get_all(self, count=-1, sort='', start=0, view='', fields='', filter='', name_prefix='', category=[]):
         """
-        Set all the labels for a resource.
-
+        Gets all items according with the given arguments.
         Args:
-            resource: The object containing the resource URI and a list of labels
-
+            start: The first item to return, using 0-based indexing.
+                If not specified, the default is 0 - start with the first available item.
+            count: The number of resources to return. A count of -1 requests all items (default).
+            sort: The sort order of the returned data set. By default, the sort order is based on create time with the
+                oldest entry first.
+            view:
+                Returns a specific subset of the attributes of the resource or collection by specifying the name of a
+                predefined view. The default view is expand (show all attributes of the resource and all elements of
+                the collections or resources).
+            fields:
+                Name of the fields.
         Returns:
-            dict: Resource Labels
+             list: A list of items matching the specified filter.
         """
-        uri = self.URI + self.RESOURCES_PATH
-        return self._client.create(resource=resource, uri=uri)
-
-    def update(self, resource):
-        """
-        Set all the labels for a resource.
-
-        Args:
-            resource (dict): Object to update.
-
-        Returns:
-            dict: Resource Labels
-        """
-        return self._client.update(resource=resource)
-
-    def delete(self, resource, timeout=-1):
-        """
-        Delete all the labels for a resource.
-
-        Args:
-            resource (dict): Object to delete.
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView, just stop waiting for its completion.
-        """
-        self._client.delete(resource=resource, timeout=timeout)
+        return self._helper.get_all(count=count, sort=sort, start=start, view=view, fields=fields, filter=filter, name_prefix=name_prefix, category=category)